py-minio is no longer used by KartonBackend
